+++ src/main/java/com/Challenge/Alkemy/service/impl/PersonajeServiceImpl.java
@@ -1,9 +1,11 @@
 package com.Challenge.Alkemy.service.impl;
 
+import com.Challenge.Alkemy.auxiliary.ErrorMessage;
 import com.Challenge.Alkemy.dto.PersonajeBasicDTO;
 import com.Challenge.Alkemy.dto.PersonajeDTO;
 import com.Challenge.Alkemy.dto.PersonajeFilterDTO;
 import com.Challenge.Alkemy.entity.Personaje;
+import com.Challenge.Alkemy.exception.ParamNotFound;
 import com.Challenge.Alkemy.mapper.PersonajeMapper;
 import com.Challenge.Alkemy.repository.PersonajeRepository;
 import com.Challenge.Alkemy.repository.specifications.PersonajeSpecification;
@@ -29,7 +31,7 @@ public class PersonajeServiceImpl implements PersonajeService {
     public PersonajeDTO findById(Long id) {
         Personaje entity= personajeRepository.findById(id).orElse(null);
         if(entity==null){
-            throw new RuntimeException("Personaje no encontrado");
+            throw new ParamNotFound(String.format(ErrorMessage.OBJECT_DOESNT_EXIST,"Personaje"));
         }
         PersonajeDTO dto=personajeMapper.personajeEntity2PersonajeDTO(entity,true);
         return dto;
@@ -37,10 +39,15 @@ public class PersonajeServiceImpl implements PersonajeService {
 
     @Override
     public void delete(Long id) {
+        Personaje entity= personajeRepository.findById(id).orElse(null);
+        if(entity==null){
+            throw new ParamNotFound(String.format(ErrorMessage.OBJECT_DOESNT_EXIST,"Personaje"));
+        }
 
         personajeRepository.deleteById(id);
     }
 
+    /* No pude hacerlo funcionar*/
     @Override
     public List<PersonajeDTO> getByFilters(String name, int age, long idMovie) {
         PersonajeFilterDTO filtersDTO = new PersonajeFilterDTO(name,age,idMovie);
@@ -70,8 +77,9 @@ public class PersonajeServiceImpl implements PersonajeService {
     public PersonajeDTO update(Long id, PersonajeDTO dto) {
         Personaje entitySearched= personajeRepository.findById(id).orElse(null);
 
+
         if(entitySearched==null){
-            throw new RuntimeException("Personaje no encontrado");
+            throw new ParamNotFound(String.format(ErrorMessage.OBJECT_DOESNT_EXIST,"Personaje"));
         }
         this.personajeMapper.personajeEntityRefreshValues(entitySearched,dto);
         Personaje entityModified=personajeRepository.save(entitySearched);
